New home page
Upon testing, i noticed that it was hard for the user to navigate and understand the point of the website. So i put the calculator into the homepage. It might seem like a good idea to put other calculators (BMI, other calorie calculators, etc) onto the navigation page, and then remove/replace some of the other tabs